requirements.txt
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.0.0
kaleido>=0.2.1  # Necessario per plotly.write_image
python-dateutil>=2.8.0
PyYAML>=6.0